refactor(core): Separate application and platform files even more
This separates application and platform code into even more files. This now removes
the ciruclar dependency between scheduling and application ref.
