+++ packages/core/src/application/application_config.ts
@@ -7,7 +7,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {EnvironmentProviders, Provider} from './di';
+import {EnvironmentProviders, Provider} from '../di';
 
 /**
  * Set of config options available during the application bootstrap operation.
