+++ packages/core/src/application/application_init.ts
@@ -8,9 +8,9 @@
 
 import {Observable} from 'rxjs';
 
-import {inject, Injectable, InjectionToken} from './di';
-import {RuntimeError, RuntimeErrorCode} from './errors';
-import {isPromise, isSubscribable} from './util/lang';
+import {inject, Injectable, InjectionToken} from '../di';
+import {RuntimeError, RuntimeErrorCode} from '../errors';
+import {isPromise, isSubscribable} from '../util/lang';
 
 /**
  * A [DI token](guide/glossary#di-token "DI token definition") that you can use to provide
